Fix logic error
The hard-coded `person` and `organization` data is index-2-based.
So to compute the next one you need to do:
    `length + 2`

I don't want to make them index-1-based right now because there is
a lot of cross-links that would need to be updated, and I feel that
is too error-prone right now.