Change linked data providers to return thunks
Previously, the linked data providers would return either the data or null.
However, this meant that determining whether a url had any linked data
about it was async.

This commit changes it to return either null or an async thunk. That way,
we can deteremine if a URL has linked data about it syncronously,
and so the "linked data" viewer will
only show up for urls that actually
have linked data, instead of all of them.

This improves the UI when selecting a parent dataset, that has no linked data.
Previously, the data browser would show an empty linked  data viewer for it.
Now, it will not and instead look at its children for linked data.